新功能开发完毕
新功能开发完毕

+++ app/src/main/java/com/sd/vr/education/presenter/FilesManager.java
@@ -2,13 +2,16 @@ package com.sd.vr.education.presenter;
 
 import java.io.File;
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.Iterator;
 import java.util.List;
+import java.util.Map;
 
 import android.os.Environment;
 import android.util.Log;
 
 import com.sd.vr.education.entity.FileDownLoad;
+import com.sd.vr.education.entity.VideoItem;
 import com.sd.vr.education.network.http.downloader.LoaderError;
 import com.sd.vr.education.network.http.downloader.LoaderExecutor;
 import com.sd.vr.education.network.http.downloader.LoaderListener;
@@ -26,10 +29,18 @@ public class FilesManager {
     public static final String DIRECTORY = Environment.getExternalStorageDirectory().getAbsolutePath()+"/com.sd.vr";
     public static final String PATCH_SUFFIX = ".mp4";
     public static final String TEMP_SUFFIX = ".cfg";
+    public static final String FILE_SUFFIX = "_";
 
-    public List<FileDownLoad> downLoadFiles = new ArrayList<>();
+    //下载状态定义
+    public static final int STATUS_TO_DOWNLOAD = 0;//待下载
+    public static final int STATUS_DOWNLOADING = 1;//下载中
+    public static final int STATUS_ERROR_DOWNLOAD = 2;//下载异常
+    public static final int STATUS_COMPLETE_DOWNLOAD = 3;//下载完成
+
+
+
+    public HashMap<FileDownLoad,Integer> downLoadFiles = new HashMap<>();
     private static FilesManager mFilesManager;
-    private String fileDownLoading;
 
     private FilesManager(){
         Log.i(TAG, "启动文件管理");
@@ -48,21 +59,42 @@ public class FilesManager {
      * @param fileDownLoad
      */
     public void downLoad(FileDownLoad fileDownLoad){
-        if (fileDownLoad == null || fileDownLoad.fileUrl.equals("") || fileDownLoad.fileName.equals("") || fileDownLoad.fileSize == 0){
+        if (fileDownLoad == null || fileDownLoad.fileUrl.equals("") || fileDownLoad.fileId.equals("") || fileDownLoad.fileSize == 0){
             return;
         }
         if (downLoadFiles == null){
-            downLoadFiles = new ArrayList<>();
+            downLoadFiles = new HashMap<>();
         }
         //如果已存在次下载项不再下载
         if (downLoadFiles.size() > 0){
-            for (FileDownLoad file:downLoadFiles) {
-                if (file.fileName.equals(fileDownLoad.fileName)){
+            for (Map.Entry<FileDownLoad, Integer> entry : downLoadFiles.entrySet()) {
+                FileDownLoad file = entry.getKey();
+                if (file.fileId.equals(fileDownLoad.fileId)){
                     return;
                 }
             }
         }
-        downLoadFiles.add(fileDownLoad);
+        //校验本地是否存在这个文件,若文件已存在但是大小校验不通过，需首先删除本地文件
+        String fileName = fileDownLoad.fileId + FILE_SUFFIX + fileDownLoad.fileNameShow;
+        long size = fileDownLoad.fileSize;
+        File fileDir = new File(DIRECTORY);
+        if (fileDir != null && fileDir.listFiles() != null && fileDir.listFiles().length > 0){
+            for (File file : fileDir.listFiles()) {
+                if (file.getAbsolutePath().endsWith(PATCH_SUFFIX)){
+                    if (file.getName().equals(fileName) && Utils.getFileSize(file) == size){
+                        //文件已下载
+                        ServiceManager.getInstance().sendDownloadAck(fileName);
+                        return;
+                    }else if (file.getName().equals(fileName) && Utils.getFileSize(file) > size){
+                        deleteFile(fileName);
+                    }
+                }
+            }
+        }
+
+        downLoadFiles.put(fileDownLoad,STATUS_TO_DOWNLOAD);
+        //刷新UI
+        ServiceManager.getInstance().updateUI();
         startDownLoad();
     }
 
@@ -89,10 +121,15 @@ public class FilesManager {
         }
 
         //TO-DO具体的删除逻辑
-        File file = new File(DIRECTORY +"/"+ fileId);
-        if (file.delete()){
-            Log.e(TAG, "delete:"+fileId);
+        File fileDir = new File(DIRECTORY);
+        for (File file : fileDir.listFiles()) {
+            if (file.getAbsolutePath().contains(fileId)){
+                if (file.delete()){
+                    Log.e(TAG, "delete:"+fileId);
+                }
+            }
         }
+
     }
 
     /**
@@ -114,21 +151,33 @@ public class FilesManager {
      */
     public void startDownLoad(){
 
-        if (fileDownLoading != null){
-            return;
-        }
-
         if (downLoadFiles == null || downLoadFiles.size() == 0){
             return;
         }
 
-        final String fileName = downLoadFiles.get(0).fileName;
-        Log.e(TAG, "待下载的内容："+fileName);
-        String url = downLoadFiles.get(0).fileUrl;
-        long size = downLoadFiles.get(0).fileSize;
+        FileDownLoad fileDownLoad = null;
+        boolean isDownLoading = false;
+        for (Map.Entry<FileDownLoad, Integer> entry : downLoadFiles.entrySet()) {
+            FileDownLoad file = entry.getKey();
+            int status = entry.getValue();
+            if (status == STATUS_DOWNLOADING){
+                isDownLoading = true;
+            }
+
+            if (status == STATUS_TO_DOWNLOAD && fileDownLoad == null){
+                fileDownLoad = file;
+            }
+        }
+
+        if (isDownLoading){
+            return;
+        }
+
+        final String fileName = fileDownLoad.fileId + FILE_SUFFIX + fileDownLoad.fileNameShow;
+        String url = fileDownLoad.fileUrl;
 
         //校验本地是否存在这个文件,若文件已存在但是大小校验不通过，需首先删除本地文件
-        File fileDir = new File(DIRECTORY);
+        /*File fileDir = new File(DIRECTORY);
         if (fileDir != null && fileDir.listFiles() != null && fileDir.listFiles().length > 0){
             for (File file : fileDir.listFiles()) {
                 if (file.getAbsolutePath().endsWith(PATCH_SUFFIX)){
@@ -141,14 +190,18 @@ public class FilesManager {
                     }
                 }
             }
-        }
-        Log.e(TAG, url);
+        }*/
+        Log.e(TAG, "待下载的内容："+url);
 
 
         LoaderInfo loaderInfo = (new LoaderInfo.Builder()).dir(DIRECTORY).name(fileName).url(url).splitter(5).build();
+        final FileDownLoad finalFileDownLoad = fileDownLoad;
+        downLoadFiles.put(finalFileDownLoad,STATUS_DOWNLOADING);//正在下载此文件
         LoaderExecutor.load(loaderInfo, new LoaderListener() {
             @Override
             public void onLoad(String url, int size, int totalSize) {
+                //刷新UI
+                ServiceManager.getInstance().updateUI();
                 Log.e(TAG, "开始,下载......");
                 Log.e(TAG, "总大小："+totalSize+"下载大小:"+size+"下载进度："+((float)size/(float) totalSize));
                 ServiceManager.getInstance().updateprocess("总大小："+totalSize+"下载大小:"+size+"下载进度："+((float)size/(float) totalSize));
@@ -159,9 +212,8 @@ public class FilesManager {
                 super.onCompleted(url, path);
                 Log.e(TAG, "完成,下载......");
                 ServiceManager.getInstance().updateprocess("完成,下载......");
-                fileDownLoading = null;
                 //移除下载项
-                clearItem(url);
+                downLoadFiles.remove(finalFileDownLoad);
                 startDownLoad();
                 //下载完成,发送下载完成指令
                 ServiceManager.getInstance().sendDownloadAck(fileName);
@@ -173,9 +225,8 @@ public class FilesManager {
             public void onCancelled(String url) {
                 super.onCancelled(url);
                 Log.e(TAG, "取消,下载......"+url);
-                fileDownLoading = null;
                 //移除下载项
-                clearItem(url);
+                downLoadFiles.remove(finalFileDownLoad);
                 //开始下一项
                 startDownLoad();
             }
@@ -185,9 +236,8 @@ public class FilesManager {
                 super.onError(url, err);
                 Log.e(TAG, "异常,下载......"+err.getMsg()+"----错误码:"+err.getCode());
                 ServiceManager.getInstance().updateprocess("异常,下载......");
-                fileDownLoading = null;
-                //移除下载项
-                clearItem(url);
+                //更新状态
+                downLoadFiles.put(finalFileDownLoad,STATUS_ERROR_DOWNLOAD);
                 //开始下一项
                 startDownLoad();
             }
@@ -199,22 +249,22 @@ public class FilesManager {
      * @param url
      */
     private void clearItem(String url){
-        if (url == null || url.equals("") ){
-            return;
-        }
-        if (downLoadFiles != null && downLoadFiles.size() > 0){
-            Iterator<FileDownLoad> sListIterator = downLoadFiles.iterator();
-            while(sListIterator.hasNext()){
-                FileDownLoad file = sListIterator.next();
-                if(url.equals(file.fileUrl)){
-                    Log.e(TAG, "移除下载任务:"+file.fileUrl);
-                    sListIterator.remove();
-                }
-            }
-        }
+//        if (url == null || url.equals("") ){
+//            return;
+//        }
+//        if (downLoadFiles != null && downLoadFiles.size() > 0){
+//            Iterator<FileDownLoad> sListIterator = downLoadFiles.iterator();
+//            while(sListIterator.hasNext()){
+//                FileDownLoad file = sListIterator.next();
+//                if(url.equals(file.fileUrl)){
+//                    Log.e(TAG, "移除下载任务:"+file.fileUrl);
+//                    sListIterator.remove();
+//                }
+//            }
+//        }
     }
 
-    public List<String> getVideoFiles(){
+    public List<VideoItem> getVideoFiles(){
         List<String> fileList = new ArrayList<>();
         List<String> downLoading = new ArrayList<>();
         File fileDir = new File(DIRECTORY);
@@ -242,7 +292,54 @@ public class FilesManager {
             }
         }
 
-        return list;
+        //搞完本地文件
+        List<VideoItem> items = new ArrayList<>();
+        for (String s:list) {
+            String[] temp = s.split(FILE_SUFFIX);
+            VideoItem item = new VideoItem();
+            item.fileId = temp[0];
+            item.fileName = s;
+            item.fileNameShow = temp[1];
+            item.fileStatus = STATUS_COMPLETE_DOWNLOAD;
+            items.add(item);
+        }
+
+        //正在下载的文件搞进来
+        for (Map.Entry<FileDownLoad, Integer> entry : downLoadFiles.entrySet()) {
+            FileDownLoad file = entry.getKey();
+            VideoItem item = new VideoItem();
+            item.fileId = file.fileId;
+            item.fileNameShow = file.fileNameShow;
+            item.fileName = file.fileId + FILE_SUFFIX + file.fileNameShow;
+            item.fileStatus = entry.getValue();
+            if (item.fileStatus == STATUS_DOWNLOADING){//计算进度
+                long size = Utils.getFileSize(new File(DIRECTORY + "/" + item.fileName));
+                float progress = size*100 / file.fileSize;
+                float num = (float)Math.round(progress*10)/10;
+                item.progress = num;
+            }
+            items.add(item);
+        }
+
+        return items;
+    }
+
+
+    /**
+     * 删除下载任务
+     * @param fileId
+     */
+    public void deteTask(String fileId){
+        Iterator<Map.Entry<FileDownLoad, Integer>> iter = downLoadFiles.entrySet().iterator();
+        while (iter.hasNext()) {
+            Map.Entry<FileDownLoad, Integer> entry =  iter.next();
+            FileDownLoad file = entry.getKey();
+            if (file.fileId.equals(fileId)){
+                iter.remove();
+                //刷新UI
+                ServiceManager.getInstance().updateUI();
+            }
+        }
     }
